fix: storage 모듈 패키지 경로 변경
