+++ storage/document/src/main/kotlin/com/gc/storage/document/group_event/GroupEventDocument.kt
@@ -1,25 +1,20 @@
-package com.gc.document.event
+package com.gc.storage.document.group_event
 
-import com.gc.document.common.BaseTimeEntity
+import com.gc.storage.document.common.BaseTimeEntity
 import jakarta.persistence.Entity
 import jakarta.persistence.Id
 import org.springframework.data.mongodb.core.mapping.Document
 import java.time.LocalDate
 import java.time.LocalTime
 
-@Document("events")
+@Document("group_events")
 @Entity
-data class EventDocument(
-  val title: String,
+data class GroupEventDocument(
+  val groupPromiseId: String,
+  val isAllDay: Boolean,
   val date: LocalDate,
   val startTime: LocalTime?,
   val endTime: LocalTime?,
-  val isAllDay: Boolean,
-  val labelId: String,
-  val alarm: String,
-  val memberId: String,
   @Id
   val id: String? = null,
-): BaseTimeEntity() {
-
-}
+): BaseTimeEntity()
