+++ storage/document/src/main/kotlin/com/gc/storage/document/share_event/MemberShareEventDocument.kt
@@ -1,6 +1,6 @@
-package com.gc.document.share_event
+package com.gc.storage.document.share_event
 
-import com.gc.document.common.BaseTimeEntity
+import com.gc.storage.document.common.BaseTimeEntity
 import jakarta.persistence.Entity
 import jakarta.persistence.Id
 import org.springframework.data.mongodb.core.mapping.Document
